音声の開始位置保存と再生時の音ズレを修正
- AudioEntt::offset_sec_をdoubleからfloatに変更(pygen非対応で保存されていなかった)
- 音声の終了位置をfend_を含めて計算(表示より1フレーム早く切れていた)
- タイムラインの波形描画に開始位置と再生速度を反映
- 再生時のフレーム進行で端数を持ち越し、音声に対して映像が遅れていく問題を修正

+++ movutl/asset/audio.hpp
@@ -14,11 +14,11 @@ public:
   AudioEntt(const char* path);
   ~AudioEntt() = default;
 
-  double offset_sec_ = 0.0;    // MPROPERTY(name="開始位置(秒)", min=0.0)
-  float speed        = 100.0f; // MPROPERTY(name="再生速度", min=0.0, max=1000.0, step=5.0)
-  float volume_      = 100.0f; // MPROPERTY(name="音量", min=0.0, max=200.0)
-  bool loop_         = false;  // MPROPERTY(name="ループ再生")
-  std::string path_;           // MPROPERTY(name="ファイル", type="path")
+  float offset_sec_ = 0.0f;   // MPROPERTY(name="開始位置(秒)", min=0.0)
+  float speed       = 100.0f; // MPROPERTY(name="再生速度", min=0.0, max=1000.0, step=5.0)
+  float volume_     = 100.0f; // MPROPERTY(name="音量", min=0.0, max=200.0)
+  bool loop_        = false;  // MPROPERTY(name="ループ再生")
+  std::string path_;          // MPROPERTY(name="ファイル", type="path")
 
   static Ref<AudioEntt> Create(const char* name, const char* path = nullptr);
   bool load_file(const char* path);
